fix: benefitDetails 한도 적용 및 타입 일관성 수정 (#5)

+++ src/main/java/org/example/cardfit/recommendation/dto/BenefitDetail.java
@@ -2,6 +2,6 @@ package org.example.cardfit.recommendation.dto;
 
 public record BenefitDetail(
         String categoryName,
-        Integer discountAmount
+        Long discountAmount
 ) {
 }
